Possibly at the point where we can send audio into a channel now.

+++ src/main/java/com/github/manevolent/ts3j/protocol/packet/PacketBody0Voice.java
@@ -57,7 +57,7 @@ public class PacketBody0Voice extends PacketBody {
         if (getRole() == ProtocolRole.SERVER)
             clientId = buffer.getShort() & 0x0000FFFF;
 
-        codecType = CodecType.fromId((int) (buffer.get() & 0xFF));
+        codecType = CodecType.fromId(buffer.get() & 0xFF);
 
         codecData = new byte[buffer.remaining()];
